Rename send form component to Recipient
