+++ src/components/send/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import connect from 'redux-connect-decorator';
 import MultiStep from '../multiStep';
-import Form from './form';
+import Form from './recipient';
 import Overview from './overview';
 import Confirm from './confirm';
 import Result from './result';
