+++ src/components/send/recipient/index.js
@@ -1,15 +1,11 @@
 import React from 'react';
 import { View } from 'react-native';
 import connect from 'redux-connect-decorator';
-import { TextEncoder } from 'text-encoding';
 import { IconButton } from '../../toolBox/button';
-import { fromRawLsk } from '../../../utilities/conversions';
-import transactions from '../../../constants/transactions';
-import { P, H1, H2, Small } from '../../toolBox/typography';
+import { P, H1 } from '../../toolBox/typography';
 import Icon from '../../toolBox/icon';
 import reg from '../../../constants/regex';
 import Input from '../../toolBox/input';
-import FormattedNumber from '../../formattedNumber';
 import { colors } from '../../../constants/styleGuide';
 import Avatar from '../../avatar';
 import Scanner from './scanner';
@@ -20,24 +16,12 @@ import getStyles from './styles';
 @connect(state => ({
   account: state.accounts.active,
 }))
-class SendForm extends React.Component {
-  references = [];
+class Recipient extends React.Component {
   activeInputRef = null;
   validator = {
     address: (str) => {
       if (str === '') return -1;
       return reg.address.test(str) ? 0 : 1;
-    },
-    amount: (str) => {
-      const { account } = this.props;
-      if (str === '') return -1;
-      return (reg.amount.test(str) &&
-        account && account.balance > transactions.send.fee &&
-        parseFloat(str) <= fromRawLsk(account.balance - transactions.send.fee)) ? 0 : 1;
-    },
-    reference: (str) => {
-      const uint8array = new TextEncoder().encode(str);
-      return uint8array.length > 64 ? 1 : 0;
     },
   };
   state = {
@@ -53,7 +37,6 @@ class SendForm extends React.Component {
       value: '',
       validity: 0,
     },
-    secondaryButtonOpacity: 1,
     avatarPreview: true,
   };
 
@@ -88,21 +71,13 @@ class SendForm extends React.Component {
     }
   }
 
-  setFormState = ({ address = '', amount = '', reference = '' }) => {
+  setFormState = ({ address = '' }) => {
     clearTimeout(this.avatarPreviewTimeout);
 
     this.setState({
       address: {
         value: address,
         validity: this.validator.address(address),
-      },
-      amount: {
-        value: amount,
-        validity: this.validator.amount(amount),
-      },
-      reference: {
-        value: reference,
-        validity: this.validator.reference(reference),
       },
       avatarPreview: false,
     });
@@ -116,19 +91,6 @@ class SendForm extends React.Component {
         avatarPreview: true,
       });
     }, 300);
-  }
-
-  /**
-   * @param {String} name - the key to set on state
-   * @param {Any} value the Value corresponding the given key
-   */
-  setReference = (value) => {
-    this.setState({
-      reference: {
-        value,
-        validity: this.validator.reference(value),
-      },
-    });
   }
 
   setAddress = (value) => {
@@ -145,38 +107,11 @@ class SendForm extends React.Component {
     this.setAvatarPreviewTimeout();
   }
 
-  setAmount = (value) => {
-    const normalizedValue = value.replace(/[^0-9]/g, '.');
-    this.setState({
-      amount: {
-        value,
-        validity: this.validator.amount(normalizedValue),
-      },
-    });
-  }
-
-  changeButtonOpacity = (val) => {
-    this.setState({ secondaryButtonOpacity: val });
-  }
-
-  changeInputFocus = (direction = 1) => {
-    let focusingRef = this.activeInputRef + direction;
-    if (focusingRef < 0) {
-      focusingRef = 0;
-    }
-    if (focusingRef > 2) {
-      focusingRef = 2;
-    }
-    this.references[`${focusingRef}`].focus();
-  }
-
   forward = () => {
     const { secondPublicKey } = this.props.account;
     const to = secondPublicKey ? 1 : 2;
     const stepData = {
-      amount: this.state.amount.value,
       address: this.state.address.value,
-      reference: this.state.reference.value,
     };
 
     this.props.move({ to, stepData });
@@ -184,10 +119,10 @@ class SendForm extends React.Component {
 
   render() {
     const {
-      navigation, account, theme, styles,
+      navigation, theme, styles,
     } = this.props;
     const {
-      address, amount, reference, avatarPreview,
+      address, avatarPreview,
     } = this.state;
 
     return (
@@ -199,7 +134,7 @@ class SendForm extends React.Component {
           setAmount={this.setAmount}
         />
         <KeyboardAwareScrollView
-            disabled={address.validity !== 0 || amount.validity !== 0 || reference.validity !== 0}
+            disabled={address.validity !== 0}
             onSubmit={this.forward}
             hasTabBar={true}
             button={{
@@ -210,20 +145,10 @@ class SendForm extends React.Component {
           >
           <View style={styles.titleContainer}>
             <View style={styles.headings}>
-              <H1 style={[styles.title, styles.theme.title]}>Send</H1>
+              <H1 style={[styles.title, styles.theme.title]}>Recipient</H1>
               <P style={[styles.subtitle, styles.theme.subtitle]}>
-                Send LSK tokens to other accounts.
+              Insert address or search a bookmark.
               </P>
-            </View>
-            <View style={[styles.balanceWrapper, styles.theme.balanceWrapper]}>
-              <Small style={[styles.subtitle, styles.theme.subtitle]}>YOUR CURRENT BALANCE</Small>
-              <View style={styles.balanceValue}>
-                <H2 style={[styles.number, styles.theme.number]}>
-                  <FormattedNumber>
-                    {fromRawLsk(account ? account.balance : 0)}
-                  </FormattedNumber>
-                </H2>
-              </View>
             </View>
           </View>
           <View style={styles.form}>
@@ -254,7 +179,7 @@ class SendForm extends React.Component {
               <Input
                 label='Address'
                 autoCorrect={false}
-                reference={(input) => { this.references[0] = input; }}
+                reference={(input) => { this.references = input; }}
                 innerStyles={{
                   errorMessage: styles.errorMessage,
                   input: [
@@ -272,34 +197,6 @@ class SendForm extends React.Component {
                 onFocus={() => { this.activeInputRef = 0; }}
               />
             </View>
-            <Input
-              label='Amount (Ⱡ)'
-              autoCorrect={false}
-              reference={(input) => { this.references[1] = input; }}
-              innerStyles={{ input: styles.input }}
-              onChange={this.setAmount}
-              value={amount.value}
-              keyboardType='numeric'
-              error={
-                amount.validity === 1 ?
-                  'Invalid amount value' : ''
-              }
-              onFocus={() => { this.activeInputRef = 1; }}
-            />
-            <Input
-              label='Reference (Optional)'
-              autoCorrect={false}
-              reference={(input) => { this.references[2] = input; }}
-              innerStyles={{ errorMessage: styles.errorMessage, input: styles.input }}
-              multiline={true}
-              onChange={this.setReference}
-              value={reference.value}
-              error={
-                reference.validity === 1 ?
-                  'Maximum length of 64 characters is exceeded.' : ''
-              }
-              onFocus={() => { this.activeInputRef = 2; }}
-            />
           </View>
         </KeyboardAwareScrollView>
       </View>
@@ -307,4 +204,4 @@ class SendForm extends React.Component {
   }
 }
 
-export default withTheme(SendForm, getStyles());
+export default withTheme(Recipient, getStyles());
